[refact] : Divide Extension code to file
