+++ RelaxOn/Extension/Color+Extension.swift
@@ -1,0 +1,21 @@
+//
+//  Color+Extensin.swift
+//  RelaxOn
+//
+//  Created by Moon Jongseek on 2022/08/25.
+//
+
+import SwiftUI
+
+extension Color {
+    // Main Color
+    static let relaxBlack = Color("RelaxBlack")
+    static let relaxRealBlack = Color("RelaxRealBlack")
+    static let relaxNightBlue = Color("RelaxNightBlue")
+    static let relaxLavender = Color("RelaxLavender")
+    static let relaxDimPurple = Color("RelaxDimPurple")
+    // System Color
+    static let systemGrey1 = Color("SystemGrey1")
+    static let systemGrey2 = Color("SystemGrey2")
+    static let systemGrey3 = Color("SystemGrey3")
+}
