+++ RelaxOn/Extension/UIColor+Extension.swift
@@ -1,0 +1,16 @@
+//
+//  UIColor+Extension.swift
+//  RelaxOn
+//
+//  Created by Moon Jongseek on 2022/08/25.
+//
+
+import UIKit
+
+extension UIColor {
+    static let relaxBlack = UIColor(named: "RelaxBlack")
+    static let relaxRealBlack = UIColor(named: "RelaxRealBlack")
+    static let relaxNightBlue = UIColor(named: "RelaxNightBlue")
+    static let relaxLavender = UIColor(named: "RelaxLavender")
+    static let relaxDimPurple = UIColor(named: "RelaxDimPurple")
+}
