+++ RelaxOn/Model/MixedSound.swift
@@ -8,18 +8,6 @@
 import Foundation
 
 struct MixedSound: Identifiable, Codable, Equatable {
-    static func == (lhs: MixedSound, rhs: MixedSound) -> Bool {
-        return true
-    }
-    
-    /// 마지막으로 생성된 id + 1 값을 반환
-    static func getUniqueId() -> Int {
-        let lastMusicId = UserDefaultsManager.shared.standard.integer(forKey: UserDefaultsManager.shared.lastMusicId)
-        let returnId = lastMusicId + 1
-        UserDefaultsManager.shared.standard.set(returnId, forKey: UserDefaultsManager.shared.lastMusicId)
-        return returnId
-    }
-
     let id: Int
     let name: String
     var baseSound: Sound?
@@ -36,5 +24,19 @@ struct MixedSound: Identifiable, Codable, Equatable {
         self.whiteNoiseSound = whiteNoiseSound
         self.imageName = imageName
         self.url = WidgetManager.getURL(id: id)
+    }
+}
+
+extension MixedSound {
+    static func == (lhs: MixedSound, rhs: MixedSound) -> Bool {
+        return true
+    }
+    
+    /// 마지막으로 생성된 id + 1 값을 반환
+    static func getUniqueId() -> Int {
+        let lastMusicId = UserDefaultsManager.shared.standard.integer(forKey: UserDefaultsManager.shared.lastMusicId)
+        let returnId = lastMusicId + 1
+        UserDefaultsManager.shared.standard.set(returnId, forKey: UserDefaultsManager.shared.lastMusicId)
+        return returnId
     }
 }
